Added respective restaurants images and description, Added 5 restaurant card containers, Embeded scrollable UI function
