+++ Capstone/app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:name=".SearchActivity"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".MainActivity_Myat"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
